backend/mac/requirements.txt
flask
flask-cors
transformers
Pillow
openai
python-dotenv
torch
torchvision
torchaudio